requirements.txt
matplotlib==3.5.3
pandas==1.3.5
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
torchsummary==1.5.1
Pillow~=9.5.0
numpy~=1.21.6